Fixed QC Checklist Items
1) Currently there are two 4-way junctions: both below each MSGEQ7
symbol, each 4-way needs to split up into two 3-way junctions.

(2) On the R3 symbol the VCC net crosses over the RESET net. It would be
nice to reorganize these nets to eliminate the intersection all
together.

(3) The Audio jacks are currently named Audiojack1 and 2, these should
have a name that uses the proper connector prefix (JPx) x being a number

(4) Two vias have tStop and bStop polygons set over them: the first is
right next to the OSHW silk, and the other is below U2 (the upper MSGEQ7
IC). These are not used for testing

(5) Audio jacks need a name and value in BRD

(6) Need measures that include the standoffs

(7) As of now the name of the product is not on the board; the previous
rev was without a silk for the product name as well (figured I would
bring it up). Also, it would be nice to get an SFE flame on the top of
the board, like the previous rev.

that's it